requirements.txt
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.0.8
charset-normalizer==3.1.0
click==8.1.8
colorama==0.4.6
decli==0.6.1
exceptiongroup==1.2.2
fastapi==0.115.6
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
importlib-metadata==6.7.0
psycopg2-binary==2.9.10
pydantic==2.10.5
pydantic_core==2.27.2
python-dotenv==1.0.1
PyYAML==6.0.2
questionary==1.10.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
termcolor==2.3.0
typing_extensions==4.12.2
uvicorn==0.34.0
watchfiles==1.0.4
websockets==14.1
